tests: add products service for tests
