+++ tests/federatedGateway/services/products.js
@@ -4,32 +4,48 @@ const require = createRequire(import.meta.url);
 const { gql } = require("apollo-server");
 
 const typeDefs = gql`
-  extend type Product @key(fields: "upc") {
-    upc: String! @external
-    weight: Int @external
-    price: Int @external
-    inStock: Boolean
-    shippingEstimate: Int @requires(fields: "price weight")
+  extend type Query {
+    topProducts(first: Int = 5): [Product]
+  }
+  type Product @key(fields: "upc") {
+    upc: String!
+    name: String
+    price: Int
+    weight: Int
   }
 `;
 
-let inventory = [
-  { upc: "1", inStock: true },
-  { upc: "2", inStock: false },
-  { upc: "3", inStock: true }
+const products = [
+  {
+    upc: "1",
+    name: "Table",
+    price: 899,
+    weight: 100
+  },
+  {
+    upc: "2",
+    name: "Couch",
+    price: 1299,
+    weight: 1000
+  },
+  {
+    upc: "3",
+    name: "Chair",
+    price: 54,
+    weight: 50
+  }
 ];
+
 
 const resolvers = {
   Product: {
     __resolveReference(object) {
-      return {
-        ...object,
-        ...inventory.find(product => product.upc === object.upc)
-      };
-    },
-    shippingEstimate: object => {
-      if (object.price > 1000) return 0;
-      return object.weight * 0.5;
+      return products.find(product => product.upc === object.upc);
+    }
+  },
+  Query: {
+    topProducts(_, args) {
+      return products.slice(0, args.first);
     }
   }
 };
@@ -38,3 +54,4 @@ export default {
   typeDefs,
   resolvers
 };
+
